Add sync client, fix features
