+++ Cargo.toml
@@ -19,10 +19,12 @@ include = [
 
 [features]
 default = ["async"]
-async = ["tokio"]
+async = ["dep:tokio"]
+sync = ["dep:bus"]
 
 [dependencies]
 rumqttc = "0.15.0"
 serde_json = "1.0"
 mqtt4bytes = "0.1.6"
 tokio = { version = "1.14.0", features = ["rt-multi-thread", "sync", "macros", "net", "time", "fs"], optional = true }
+bus = { version = "2.4.0", optional = true }
