+++ examples/pubsub.rs
@@ -1,7 +1,6 @@
-use serde_json::json;
+use aws_iot_device_sdk_rust::{async_event_loop_listener, AWSIoTAsyncClient, AWSIoTSettings};
 use rumqttc::{self, Packet, QoS};
 use std::error::Error;
-use aws_iot_device_sdk_rust::client::{AWSIoTSettings, AWSIoTAsyncClient, async_event_loop_listener};
 
 #[tokio::main]
 async fn main() -> Result<(), Box<dyn Error>> {
@@ -25,38 +24,31 @@ async fn main() -> Result<(), Box<dyn Error>> {
 
     let recv1_thread = tokio::spawn(async move {
         loop {
-            match receiver1.recv().await {
-                Ok(event) => {
-                    match event {
-                        Packet::Publish(p) => println!("Received message {:?} on topic: {}", p.payload, p.topic),
-                        _ => println!("Got event on receiver1: {:?}", event),
+            if let Ok(event) = receiver1.recv().await {
+                match event {
+                    Packet::Publish(p) => {
+                        println!("Received message {:?} on topic: {}", p.payload, p.topic)
                     }
-
-                },
-                Err(_) => (),
+                    _ => println!("Got event on receiver1: {:?}", event),
+                }
             }
         }
     });
 
     let recv2_thread = tokio::spawn(async move {
         loop {
-            match receiver2.recv().await {
-                Ok(event) => println!("Got event on receiver2: {:?}", event),
-                Err(_) => (),
+            if let Ok(event) = receiver2.recv().await {
+                println!("Got event on receiver2: {:?}", event);
             }
         }
     });
     let listen_thread = tokio::spawn(async move {
-            async_event_loop_listener(eventloop_stuff).await.unwrap();
-            //iot_core_client.listen().await.unwrap();
+        async_event_loop_listener(eventloop_stuff).await.unwrap();
+        //iot_core_client.listen().await.unwrap();
     });
 
     //iot_core_client.publish("topic".to_string(), QoS::AtMostOnce, "hey").await.unwrap();
-    tokio::join!(
-        recv1_thread,
-        recv2_thread,
-        listen_thread);
-        
-    Ok(())
+    tokio::join!(recv1_thread, recv2_thread, listen_thread);
 
+    Ok(())
 }
